tests: make sure we empty the queue, and keep counters thread-private until end of test.
Only check consumer exit condition when dequeue fails,
so we don't need ctx->count to be updated.

This allows us to drop that point of contention between readers and writers.

Using a local counter and only doing an atomic add to the shared variable
at the end also reduces contention between threads in the test harness itself.