dequeue: nicer var names.  Lots of commented-out experimental stuff.
remove unused int cn_runtimes = 0; and remove ctx->count updates.

Move ret declaration to where it's initialized.

The mb() in the do{}while() loop is necessary to make a seq-cst store,
but loading old_head->next twice in a row is not necessary.

Add an #ifed out version of the loop that doesn't do the HP stuff.

# Conflicts:
#	lfq.c